+++ gt7dashboard/test/test_gt7diagrams.py
@@ -13,11 +13,11 @@ from gt7dashboard.gt7diagrams import (
 from gt7dashboard.gt7lap import Lap
 from gt7dashboard.gt7racediagram import RaceDiagram
 from gt7dashboard.race_time_datatable import RaceTimeDataTable
-
+from gt7dashboard.gt7lapstorage import load_laps_from_json
 
 class TestHelper(unittest.TestCase):
     def setUp(self) -> None:
-        self.test_laps = gt7helper.load_laps_from_json("test_data/broad_bean_raceway_time_trial_4laps.json")
+        self.test_laps = load_laps_from_json("test_data/broad_bean_raceway_time_trial_4laps.json")
 
     def test_get_throttle_braking_race_line_diagram(self):
         (
